name & value

+++ iflow/common/src/main/java/com/pth/iflow/common/enums/base/IValueList.java
@@ -1,7 +1,6 @@
 package com.pth.iflow.common.enums.base;
 
-
 public interface IValueList {
-  Long getValue();
+  Integer getValue();
 
 }
